Migrate to esm
In this diff I converted source to esm so we can easily use this lib with
native browser modules. This also allow bundlers do not add additional
runtime to work around cjs and as a result produce smaller bundles.

I converted index.js to esm and produced cjs with rollup. Added
`pretest` and `prepublish` npm hooks to test and publish safely.